+++ BioMed.Api/BioMed.Api/LoginModels/LoginRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace BioMed.Api.LoginModels
+{
+    public class LoginRequest
+    {
+        public string Login { get; set; }
+        public string Password { get; set; }
+    }
+}
