+++ BioMed.Api/BioMed.Api/LoginModels/User.cs
@@ -1,0 +1,10 @@
+﻿namespace BioMed.Api.LoginModels
+{
+    public class User
+    {
+        public string? Name { get; set; }
+        public string? Phone { get; set; }
+        public string? Login { get; set; }
+        public string? Password { get; set; }
+    }
+}
